requirements.txt
fastapi
uvicorn
python-multipart
Pillow
google-generativeai
langchain-google-genai
langchain-core
sentence-transformers
faiss-cpu
numpy
PyMuPDF
fpdf
pydantic
python-dotenv